// src/store/main/system/types.ts
export interface systemState {
  userList: any[]
  userCount: number
  roleList: any[]
  roleCount: number
  goodsList: any[]
  goodsCount: number
  menuList: any[]
  menuCount: number
  departmentTotalCount: number
  departmentList: any[]
  storyTotalCount: number
  storyList: any[]
  categoryTotalCount: number
  categoryList: any[]
}